+++ Engine/main/Objects/Enemy.java
@@ -18,7 +18,7 @@ public class Enemy extends GameObject {
         this.handler = handler;
     }
 
-    public Rectangle getBounds(){return new Rectangle(x,y,16,16);}
+    public Rectangle getBounds(){return new Rectangle((int)x,(int)y,16,16);}
 
     @Override
     public void tick() {
@@ -28,13 +28,13 @@ public class Enemy extends GameObject {
         if(y <= 0 || y >= Game.HEIGHT-32) velY *= -1;
         if(x <= 0 || x >= Game.WIDTH-16) velX *= -1;
 
-        handler.addObject(new Trail(x,y,ID.trail,Color.red,16,16, 0.1f,handler));
+        handler.addObject(new Trail((int)x,(int)y,ID.trail,Color.red,16,16, 0.1f,handler));
     }
 
     @Override
     public void render(Graphics g) {
         g.setColor(Color.red);
-        g.fillRect(x,y,16,16);
+        g.fillRect((int)x,(int)y,16,16);
 
     }
 }
